Move TestDirectory related code to new nokeetest library
