+++ gradle/libraries/nokeetest-libraries/test-directory-junit-jupiter/build.gradle
@@ -1,0 +1,12 @@
+plugins {
+	id 'java-library'
+}
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+	api 'org.junit.jupiter:junit-jupiter:5.8.2'
+	api project(':test-directory')
+}
