+++ gradle/libraries/nokeetest-libraries/test-directory/build.gradle
@@ -1,0 +1,13 @@
+plugins {
+	id 'java-library'
+}
+
+repositories {
+	mavenCentral()
+}
+
+dependencies {
+//	implementation 'org.apache.commons:commons-lang3:3.12.0'
+	implementation 'commons-io:commons-io:2.11.0'
+	implementation 'dev.failsafe:failsafe:3.2.3'
+}
