+++ gradle/libraries/nokeetest-libraries/test-directory/src/main/java/net/nokeedev/testing/file/TestDirectoryProvider.java
@@ -13,13 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.internal.testing.file;
+package net.nokeedev.testing.file;
 
-import java.lang.annotation.*;
+import java.nio.file.Path;
 
-@Target({ElementType.FIELD, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-@Documented
-public @interface TestDirectory {
+/**
+ * Implementations provide a working space to be used in tests.
+ *
+ * The client is not responsible for removing any files.
+ */
+public interface TestDirectoryProvider {
+
+	/**
+	 * The directory to use, guaranteed to exist.
+	 *
+	 * @return The directory to use, guaranteed to exist.
+	 */
+	Path getTestDirectory();
+
+	void suppressCleanup();
+
+	void suppressCleanupErrors();
 
 }
