+++ settings.gradle
@@ -15,6 +15,7 @@ rootProject.name = 'gradle-native'
 
 includeBuild 'subprojects/templates'
 includeBuild 'gradle/plugins/docs-gradle-plugin'
+includeBuild 'gradle/libraries/nokeetest-libraries'
 
 /**
  * Adds specified subproject path to this Gradle build.
